requirements.txt
samila
pillow
matplotlib